[TS-293] Refactor: StarMainPage 스타일 코드 분리
