+++ src/pages/StarMainPage/StarMainPage.tsx
@@ -1,5 +1,3 @@
-import { css } from "@emotion/react";
-
 import Gnb from "@/components/common/Gnb/Gnb";
 import CharacterUnlockModal from "@/components/StarPage/Modal/CharacterUnlockModal";
 import StarBackground from "@/components/StarPage/StarMain/StarBackground";
@@ -13,7 +11,7 @@ import { openModal } from "@/api/modal/modalSlice";
 
 import { modalType } from "@/constants/modalConstants";
 
-import { theme } from "@/styles/theme";
+import { containerStyle, starMainPageGnbStyle } from "@/pages/StarMainPage/StarMainPage.style";
 
 export default function StarMainPage() {
 	const dispatch = useAppDispatch();
@@ -43,48 +41,6 @@ export default function StarMainPage() {
 		</div>
 	);
 }
-
-const containerStyle = css`
-	height: 100dvh;
-	background: linear-gradient(
-		#0b0a0b 2%,
-		#192771 38%,
-		#3e53b3 63%,
-		#4166b7 76%,
-		#6b8acb 84%,
-		#6d68d4 98%
-	);
-
-	.wrapper-top {
-		width: 22.5rem;
-		padding: 1.75rem 1.5rem 0;
-		margin: 0 auto;
-	}
-
-	.wrapper-middle {
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		height: calc(100dvh - 20.825rem);
-	}
-
-	.wrapper-bottom {
-		width: 22.5rem;
-		padding: 0 1.5rem 4.75rem;
-		position: absolute;
-		bottom: 0;
-		left: 50%;
-		transform: translateX(-50%);
-	}
-`;
-
-const starMainPageGnbStyle = css`
-	opacity: 0.5;
-
-	& > div {
-		background-color: ${theme.color.font_black};
-	}
-`;
 
 // TODO: 테스트를 위한 임시 데이터(추후 삭제 예정)
 export const data = {
